requirements.txt
fastapi==0.129.2
natsort==8.4.0
nicegui==3.7.1
numpy==2.4.2
opencv_contrib_python==4.10.0.84
opencv_python==4.10.0.84
opencv_python_headless==4.11.0.86
Pillow==12.1.1
pyclipper==1.3.0.post5
PyYAML==6.0.2
PyYAML==6.0.3
Requests==2.32.5
scipy==1.17.0
setuptools==74.1.2
Shapely==2.1.2
simple_lama_inpainting==0.1.2
torch==2.6.0
torch_summary==1.4.5
torchsummary==1.5.1
torchvision==0.21.0
wandb==0.25.0
